move desktop Popup to skikoMain sourceSet
