+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/window/Popup.skiko.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2022 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.LocalComposeScene
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.awt.LocalLayerContainer
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.input.key.KeyEvent
 import androidx.compose.ui.layout.Layout
@@ -34,16 +33,12 @@ import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.SkiaBasedOwner
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.setContent
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntRect
 import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.LayoutDirection
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.round
-import java.awt.MouseInfo
-import javax.swing.SwingUtilities.convertPointFromScreen
 
 /**
  * Opens a popup with the given content.
@@ -213,68 +208,6 @@ private fun PopupLayout(
             scene.detach(owner)
             composition.dispose()
             owner.dispose()
-        }
-    }
-}
-
-/**
- * Provides [PopupPositionProvider] relative to the current mouse cursor position.
- *
- * @param offset [DpOffset] to be added to the position of the popup.
- * @param alignment The alignment of the popup relative to the current cursor position.
- * @param windowMargin Defines the area within the window that limits the placement of the popup.
- */
-@Composable
-fun rememberCursorPositionProvider(
-    offset: DpOffset = DpOffset.Zero,
-    alignment: Alignment = Alignment.BottomEnd,
-    windowMargin: Dp = 4.dp
-): PopupPositionProvider = with(LocalDensity.current) {
-    val component = LocalLayerContainer.current
-    val cursorPoint = remember {
-        val awtMousePosition = MouseInfo.getPointerInfo().location
-        convertPointFromScreen(awtMousePosition, component)
-        IntOffset(
-            (awtMousePosition.x * component.density.density).toInt(),
-            (awtMousePosition.y * component.density.density).toInt()
-        )
-    }
-    val offsetPx = IntOffset(offset.x.roundToPx(), offset.y.roundToPx())
-    val windowMarginPx = windowMargin.roundToPx()
-    object : PopupPositionProvider {
-        override fun calculatePosition(
-            anchorBounds: IntRect,
-            windowSize: IntSize,
-            layoutDirection: LayoutDirection,
-            popupContentSize: IntSize
-        ) = with(density) {
-            val anchor = IntRect(cursorPoint, IntSize.Zero)
-            val tooltipArea = IntRect(
-                IntOffset(
-                    anchor.left - popupContentSize.width,
-                    anchor.top - popupContentSize.height,
-                ),
-                IntSize(
-                    popupContentSize.width * 2,
-                    popupContentSize.height * 2
-                )
-            )
-            val position = alignment.align(popupContentSize, tooltipArea.size, layoutDirection)
-            var x = tooltipArea.left + position.x + offsetPx.x
-            var y = tooltipArea.top + position.y + offsetPx.y
-            if (x + popupContentSize.width > windowSize.width - windowMarginPx) {
-                x -= popupContentSize.width
-            }
-            if (y + popupContentSize.height > windowSize.height - windowMarginPx) {
-                y -= popupContentSize.height + anchor.height
-            }
-            if (x < windowMarginPx) {
-                x = windowMarginPx
-            }
-            if (y < windowMarginPx) {
-                y = windowMarginPx
-            }
-            IntOffset(x, y)
         }
     }
 }
